Reorganize Runtime exceptions
