+++ src/php/Command/ReplCommand.php
@@ -9,9 +9,9 @@ use Phel\Command\Repl\ReplCommandIoInterface;
 use Phel\Compiler\Analyzer\Exceptions\CompilerException;
 use Phel\Compiler\EvalCompilerInterface;
 use Phel\Compiler\Parser\Exceptions\UnfinishedParserException;
-use Phel\Exceptions\ExceptionPrinterInterface;
 use Phel\Exceptions\ExitException;
 use Phel\Printer\PrinterInterface;
+use Phel\Runtime\Exceptions\ExceptionPrinterInterface;
 use Throwable;
 
 final class ReplCommand
