+++ src/php/Runtime/Exceptions/ExceptionArgsPrinter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Exceptions\Printer;
+namespace Phel\Runtime\Exceptions;
 
 use Phel\Printer\PrinterInterface;
