+++ src/php/Runtime/Exceptions/Extractor/SourceMapExtractor.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Phel\Exceptions\Extractor;
+namespace Phel\Runtime\Exceptions\Extractor;
 
-use Phel\Exceptions\Extractor\ReadModel\SourceMapInformation;
+use Phel\Runtime\Exceptions\Extractor\ReadModel\SourceMapInformation;
 
 final class SourceMapExtractor implements SourceMapExtractorInterface
 {
