+++ src/php/Runtime/Exceptions/HtmlExceptionPrinter.php
@@ -2,14 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Phel\Exceptions;
+namespace Phel\Runtime\Exceptions;
 
 use Phel\Compiler\Emitter\OutputEmitter\Munge;
 use Phel\Compiler\Emitter\OutputEmitter\MungeInterface;
 use Phel\Compiler\Exceptions\PhelCodeException;
 use Phel\Compiler\Parser\ReadModel\CodeSnippet;
-use Phel\Exceptions\Printer\ExceptionArgsPrinter;
-use Phel\Exceptions\Printer\ExceptionArgsPrinterInterface;
 use Phel\Lang\FnInterface;
 use Phel\Printer\Printer;
 use ReflectionClass;
