+++ tests/php/Integration/Repl/ReplIntegrationTest.php
@@ -10,11 +10,11 @@ use Phel\Command\ReplCommand;
 use Phel\Compiler\Analyzer\Environment\GlobalEnvironment;
 use Phel\Compiler\CompilerFactory;
 use Phel\Compiler\Emitter\OutputEmitter\Munge;
-use Phel\Exceptions\Extractor\FilePositionExtractor;
-use Phel\Exceptions\Extractor\SourceMapExtractor;
-use Phel\Exceptions\Printer\ExceptionArgsPrinter;
-use Phel\Exceptions\TextExceptionPrinter;
 use Phel\Printer\Printer;
+use Phel\Runtime\Exceptions\ExceptionArgsPrinter;
+use Phel\Runtime\Exceptions\Extractor\FilePositionExtractor;
+use Phel\Runtime\Exceptions\Extractor\SourceMapExtractor;
+use Phel\Runtime\Exceptions\TextExceptionPrinter;
 use Phel\Runtime\RuntimeFactory;
 use PHPUnit\Framework\TestCase;
 use RecursiveDirectoryIterator;
