+++ tests/php/Unit/Exceptions/Extractor/FilePositionExtractorTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace PhelTest\Unit\Exceptions\Extractor;
 
-use Phel\Exceptions\Extractor\FilePositionExtractor;
-use Phel\Exceptions\Extractor\ReadModel\FilePosition;
-use Phel\Exceptions\Extractor\ReadModel\SourceMapInformation;
-use Phel\Exceptions\Extractor\SourceMapExtractorInterface;
+use Phel\Runtime\Exceptions\Extractor\FilePositionExtractor;
+use Phel\Runtime\Exceptions\Extractor\ReadModel\FilePosition;
+use Phel\Runtime\Exceptions\Extractor\ReadModel\SourceMapInformation;
+use Phel\Runtime\Exceptions\Extractor\SourceMapExtractorInterface;
 use PHPUnit\Framework\TestCase;
 
 final class FilePositionExtractorTest extends TestCase
